feat: summary functions
