+++ lib/functions/index.ts
@@ -1,2 +1,3 @@
+export * from "./summary";
 export * from "./buildTimeline";
 export * from "./postRequest";
